pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "adli-sdk"
version = "0.1.0"
description = "Official Python SDK for ADLI — runtime strategy learning for AI agents"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.10"
dependencies = [
    "httpx>=0.27",
    "pydantic>=2.0",
    "opentelemetry-sdk>=1.20",
    "opentelemetry-api>=1.20",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-timeout>=2.3",
    "ruff>=0.4",
    "pydantic-ai>=1.0",
    "langchain-core>=0.3",
    "langchain-community>=0.3",
    "langchain>=0.3",
    "langchain-openai>=0.2",
    "langchain-classic>=0.1",
    "langgraph>=0.2",
    "crewai>=0.30",
    "llama-index-core>=0.10",
    "openai-agents>=0.0.3",
]

[tool.hatch.build.targets.wheel]
packages = ["src/adli_sdk"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
timeout = 30
markers = ["integration: tests hitting real ADLI backend"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
